Expect double rather than in for start/end/step
Following exception was thrown for start=0.0 when converting
an ome.tiff with modulo annotation back to ome.tiff:

```
Populating metadata
Exception in thread "main" java.lang.NumberFormatException: For input string: "0.0"
    at java.lang.NumberFormatException.forInputString(NumberFormatException.java:65)
    at java.lang.Integer.parseInt(Integer.java:492)
    at java.lang.Integer.parseInt(Integer.java:527)
    at loci.formats.services.OMEXMLServiceImpl.getModuloAlong(OMEXMLServiceImpl.java:564)
    at loci.formats.services.OMEXMLServiceImpl.getModuloAlongT(OMEXMLServiceImpl.java:528)
    at loci.formats.in.OMETiffReader.initFile(OMETiffReader.java:834)
    at loci.formats.FormatReader.setId(FormatReader.java:1360)
    at loci.formats.ImageReader.setId(ImageReader.java:781)
    at loci.formats.tools.ImageConverter.testConvert(ImageConverter.java:337)
    at loci.formats.tools.ImageConverter.main(ImageConverter.java:672)
```

+++ components/scifio/src/loci/formats/services/OMEXMLServiceImpl.java
@@ -561,13 +561,13 @@ public class OMEXMLServiceImpl extends AbstractService implements OMEXMLService
           Node unit = attrs.getNamedItem("Unit");
 
           if (start != null) {
-            m.start = Integer.parseInt(start.getNodeValue());
+            m.start = Double.parseDouble(start.getNodeValue());
           }
           if (end != null) {
-            m.end = Integer.parseInt(end.getNodeValue());
+            m.end = Double.parseDouble(end.getNodeValue());
           }
           if (step != null) {
-            m.step = Integer.parseInt(step.getNodeValue());
+            m.step = Double.parseDouble(step.getNodeValue());
           }
           if (type != null) {
             m.type = type.getNodeValue();
